requirements.txt

selenium
python-dotenv
pillow
